Added basic test, need more cases
